Customize delegated authn web flow config
The "client action" action now transits to the non-interactive
authn check action when the returned event is success, which
currently just passes the flow on to the TGT creation action
w/o extra processing. In the future, it will identify whether
it is ORCID login or institution login via the CAS client and
thus handles them differently (to-be-implemented).